Added modulo operation

+++ proj01/src/task01.non_visitor/src/include/task01.non_visitor/tree_node.hpp
@@ -17,6 +17,6 @@ class TreeNode
     virtual std::string to_infix()   = 0;
     virtual std::string to_postfix() = 0;
     virtual std::string to_prefix()  = 0;
-    virtual double evaluate()        = 0;
+    virtual int evaluate()           = 0;
 };
 } // namespace psso
